+++ packages/moonstone/IncrementSlider/IncrementSlider.js
@@ -120,6 +120,24 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			onIncrement: PropTypes.func,
 
 			/**
+			 * The handler to run when the component is removed while retaining focus.
+			 *
+			 * @type {Function}
+			 * @param {Object} event
+			 * @public
+			 */
+			onSpotlightDisappear: PropTypes.func,
+
+			/**
+			 * When `true`, the component cannot be navigated using spotlight.
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			spotlightDisabled: PropTypes.bool,
+
+			/**
 			* The amount to increment or decrement the value.
 			*
 			* @type {Number}
@@ -152,6 +170,7 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			max: 100,
 			min: 0,
 			pressed: false,
+			spotlightDisabled: false,
 			step: 1,
 			value: 0,
 			vertical: false
@@ -170,20 +189,24 @@ const IncrementSliderBaseFactory = factory({css: componentCss}, ({css}) => {
 			incrementIcon: ({incrementIcon, vertical}) => (incrementIcon || (vertical ? 'arrowlargeup' : 'arrowlargeright'))
 		},
 
-		render: ({decrementDisabled, decrementIcon, incrementDisabled, incrementIcon, onIncrement, onDecrement, incrementSliderClasses, ...rest}) => (
+		render: ({decrementDisabled, decrementIcon, incrementDisabled, incrementIcon, incrementSliderClasses, onIncrement, onDecrement, onSpotlightDisappear, spotlightDisabled, ...rest}) => (
 			<div className={incrementSliderClasses}>
 				<IncrementSliderButton
 					className={css.decrementButton}
 					disabled={decrementDisabled}
 					onClick={onDecrement}
+					onSpotlightDisappear={onSpotlightDisappear}
+					spotlightDisabled={spotlightDisabled}
 				>
 					{decrementIcon}
 				</IncrementSliderButton>
-				<Slider {...rest} className={css.slider} />
+				<Slider {...rest} className={css.slider} onSpotlightDisappear={onSpotlightDisappear} spotlightDisabled={spotlightDisabled} />
 				<IncrementSliderButton
 					className={css.incrementButton}
 					disabled={incrementDisabled}
 					onClick={onIncrement}
+					onSpotlightDisappear={onSpotlightDisappear}
+					spotlightDisabled={spotlightDisabled}
 				>
 					{incrementIcon}
 				</IncrementSliderButton>