+++ packages/moonstone/IncrementSlider/IncrementSliderButton.js
@@ -31,7 +31,7 @@ const IncrementSliderButtonBase = kind({
 	)
 });
 
-const OnlyUpdate = onlyUpdateForKeys(['children', 'disabled']);
+const OnlyUpdate = onlyUpdateForKeys(['children', 'disabled', 'spotlightDisabled']);
 const IncrementSliderButton = OnlyUpdate(IncrementSliderButtonBase);
 
 export default IncrementSliderButton;